+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/subjects/movie/MovieUiState.kt
@@ -4,7 +4,8 @@ import com.github.bumblebee202111.doubean.model.Movie
 import com.github.bumblebee202111.doubean.model.SubjectWithInterest
 
 sealed interface MovieUiState {
-    data class Success(val movie: SubjectWithInterest<Movie>) : MovieUiState
+    data class Success(val movie: SubjectWithInterest<Movie>, val isLoggedIn: Boolean) :
+        MovieUiState
     data object Loading : MovieUiState
     data object Error : MovieUiState
 }